requirements.txt
chainlit==0.7.700
langchain==0.2.5
langchain-community
langchain_core==0.2.9
langchain_huggingface==0.0.3
openai==1.35.3
langchain-openai
langchain_text_splitters==0.2.1
python-dotenv==1.0.1
tiktoken==0.7.0
qdrant-client==1.9.2
pymupdf